mbm: refactor - incubation mb reservation moved to controller
